Changed solution to reject trees with duplicate values, since solution couldn't properly validate trees with duplicate values.
(1)
Our first thought might be to do an in-order traversal, copy the elements to an array, and then check to see if the array is sorted. This solution takes up a bit of extra memory, but it works -- mostly. 
The only problem is that it can’t handle duplicate values in the tree properly. For example, the algorithm cannot distinguish between the two trees below (one of which is invalid) since they have the same in-order traversal.
Valid BST [20.left = 20]   
Invalid BST [20.right = 20]

However, if we assume that the tree cannot have duplicate values, then this approach works. The pseudocode for this method looks something like:

(2) Changed line 14 of pseudocode from "if (array[i] < array[i - 1]) return false;" to "if (array[i] <= array[i - 1]) return false;"

(3) Changed line 9 of real code from "if (n.data < last_printed) return false;" to "if (n.data <= last_printed) return false;"